requirements.txt
tensorflow
pillow
numpy
Flask
Werkzeug
requests    # Good practice to include if used, even indirectly by tf.keras sometimes
gunicorn    # Optional: For production deployment